Transfers the modules into files
