+++ src/components/load-more-btn.js
@@ -1,0 +1,9 @@
+/**
+ * Returns the markup of «load more» button
+ * @return {string}
+ */
+const createLoadMoreBtnTemplate = () => (
+  `<button class="load-more" type="button">load more</button>`
+);
+
+export {createLoadMoreBtnTemplate};
